Style: Create contact thanks page design, with percentage margin and paddings
