+++ package.json
@@ -9,9 +9,9 @@
     "@reduxjs/toolkit": "^1.8.1",
     "@testing-library/jest-dom": "^5.16.4",
     "@testing-library/user-event": "^14.1.1",
-    "moment": "^2.29.4",
     "firebase": "^9.15.0",
     "firebaseui": "^6.0.2",
+    "moment": "^2.29.4",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-icons": "^4.7.1",
@@ -60,7 +60,6 @@
     "eslint-plugin-jsx-a11y": "^6.5.1",
     "eslint-plugin-react": "^7.28.0",
     "eslint-plugin-react-hooks": "^4.3.0",
-    "jest": "^27.5.1",
     "jest": "^27.5.1",
     "postcss": "^8.4.17",
     "prettier": "2.7.1",
