+++ src/components/contact/ContactForm.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function ContactForm() {
+  return <div>Hello</div>;
+}
+
+export default ContactForm;
